refactor: extract clipboard and paste logic
